requirements.txt
flask
langgraph
playwright
spacy
python-dotenv
google-genai
google-generativeai
pytest
